Artikel überarbeitung -neue query find by preis

+++ shop/src/main/java/shop/artikelverwaltung/service/ArtikelService.java
@@ -2,6 +2,7 @@ package shop.artikelverwaltung.service;
 
 
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import shop.artikelverwaltung.domain.AbstractArtikel;
@@ -11,7 +12,7 @@ public interface ArtikelService {
 	AbstractArtikel findArtikelById(Long id);
 	List<AbstractArtikel> findArtikelByIds(List<Long> ids); 
 	List<AbstractArtikel> findArtikelByName (String name);
-	List<AbstractArtikel> findArtikelByPrice (int preis);
+	List<AbstractArtikel> findArtikelByPrice (BigDecimal preis);
 	AbstractArtikel createArtikel(AbstractArtikel artikel);
 	void updateRad(Rad rad);
 }
